Fix Automerge hydration for production backups
Add custom hydration helpers to handle TypeScript/Automerge data:
- String fields stored as Text objects (collaborative editing)
- Numeric fields stored as Int instead of F64
- Optional fields that are absent (not null)

Convert timestamps and intervals from f64 to i64 to match
JavaScript Number serialization in Automerge documents.

Add golden file test with real production backup (68 tasks)
to verify hydration compatibility.

+++ crates/tasklens-core/src/domain/constants.rs
@@ -11,4 +11,4 @@ pub const FEEDBACK_SENSITIVITY: f64 = 2.0;
 pub const MIN_PRIORITY: f64 = 0.001;
 
 /// Default lead time in milliseconds (8 hours).
-pub const DEFAULT_LEAD_TIME_MILLIS: f64 = 8.0 * 60.0 * 60.0 * 1000.0;
+pub const DEFAULT_LEAD_TIME_MILLIS: i64 = 8 * 60 * 60 * 1000;
