+++ crates/tasklens-store/src/actions.rs
@@ -17,14 +17,14 @@ pub enum Action {
     },
     CompleteTask {
         id: TaskID,
-        current_time: f64,
+        current_time: i64,
     },
     MoveTask {
         id: TaskID,
         new_parent_id: Option<TaskID>,
     },
     RefreshLifecycle {
-        current_time: f64,
+        current_time: i64,
     },
 }
 
@@ -35,9 +35,9 @@ pub struct TaskUpdates {
     /// Optional update: `Some(None)` clears the place, `Some(Some(id))` sets it, `None` keeps current.
     pub place_id: Option<Option<PlaceID>>,
     /// Optional update: `Some(None)` clears the due date, `Some(Some(ts))` sets it, `None` keeps current.
-    pub due_date: Option<Option<f64>>,
+    pub due_date: Option<Option<i64>>,
     pub schedule_type: Option<tasklens_core::types::ScheduleType>,
-    pub lead_time: Option<Option<f64>>,
+    pub lead_time: Option<Option<i64>>,
     pub repeat_config: Option<Option<tasklens_core::types::RepeatConfig>>,
     pub is_sequential: Option<bool>,
 }
